requirements.txt
# # Web framework
# fastapi
# uvicorn[standard]
# pydantic

# # Google AI
# google-generativeai
# google-cloud-aiplatform
# google-cloud-documentai
# google-cloud-storage
# google-cloud-discoveryengine

# # Supabase
# supabase==2.9.1
# httpx<0.28.0

# # Utilities
# requests
# urllib3
# rich
# json-repair
# python-dotenv

# # Document generation
# reportlab
# python-pptx
# PyMuPDF
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4

# Google AI
google-generativeai==0.7.2
google-cloud-aiplatform==1.64.0
google-cloud-documentai==2.30.0
google-cloud-storage==2.17.0
google-cloud-discoveryengine==0.11.13

# Supabase
supabase==2.9.1
httpx==0.27.0

# Utilities
requests==2.32.3
urllib3==2.2.2
rich==13.7.1
json-repair==0.30.0
python-dotenv==1.0.1

# Document generation
reportlab==4.2.2
python-pptx==0.6.23
PyMuPDF==1.24.9


# Image generation
huggingface_hub==0.24.0
Pillow==10.4.0
